requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
blessed==1.20.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
contextlib2==21.6.0
contourpy==1.0.7
cosine-annealing-warmup @ git+https://github.com/katsura-jp/pytorch-cosine-annealing-with-warmup@12d03c07553aedd3d9e9155e2b3e31ce8c64081a
croniter==1.3.10
cycler==0.11.0
dateutils==0.6.12
deepdiff==6.3.0
dnspython==2.3.0
docker-pycreds==0.4.0
docstring-parser==0.15
einops==0.6.1
email-validator==1.3.1
fastapi==0.88.0
filelock==3.11.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.4
hydra-core==1.3.2
idna==3.4
importlib-resources==5.12.0
inquirer==3.1.3
ipdb==0.13.13
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonargparse==4.20.1
kiwisolver==1.4.4
lightning==2.0.1
lightning-cloud==0.5.32
lightning-utilities==0.8.0
lit==16.0.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
ml-collections==0.1.1
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
opencv-python==4.7.0.72
ordered-set==4.1.0
orjson==3.8.9
packaging==23.0
pandas==2.0.0
pathtools==0.1.2
Pillow==9.5.0
protobuf==3.20.3
pydantic==1.10.7
pydicom==2.3.1
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrootutils==1.0.4
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
readchar==4.0.5
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==13.3.3
sentry-sdk==1.19.1
-e git+ssh://git@github.com/fuying-wang/Multi-seq-mae.git@3dbbdf6a1dddfb8ac56b2980612b277c350de929#egg=seqmae
setproctitle==1.3.2
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tensorboardX==2.6
timm==0.6.13
tokenizers==0.13.3
torch==2.0.0
torchaudio==2.0.1
torchmetrics==0.11.4
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.4
triton==2.0.0
typeshed-client==2.2.0
tzdata==2023.3
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
wandb==0.14.2
watchfiles==0.19.0
websocket-client==1.5.1
websockets==11.0.1
yarl==1.8.2
